chore(kubernetes): update image pull policy and add tolerations for Longhorn Helm release
Modified the Helm release configuration for the Enshrouded application to set the image pull policy to 'Always', ensuring the latest image is always used. Additionally, added tolerations for system taints in the Longhorn Helm release to allow scheduling on nodes that are initializing, improving resource utilization and scheduling efficiency in the cluster.